requirements.txt
numpy
pandas
python-dotenv
requests
azure-ai-vision-imageanalysis
opencv-python
pyodbc
pyzbar
azure.cognitiveservices.vision.computervision
Pillow
pytest
selenium
streamlit



